✨ #7 feat : integrate typescript-obfuscator into vscode extension
