+++ engines/typescript-obfuscator/.gitignore
@@ -1,0 +1,1 @@
+**/declassify-typescript-obfuscator-*.tgz
