+++ engines/typescript-obfuscator/src/index.ts
@@ -1,3 +1,3 @@
-export { obfuscate } from './obfuscate';
-export { deobfuscate, DeobfuscationResult } from './deobfuscate';
-export type { ObfuscationResult } from './obfuscate';
+export { obfuscate } from './obfuscate.js';
+export { deobfuscate, DeobfuscationResult } from './deobfuscate.js';
+export type { ObfuscationResult } from './obfuscate.js';
