+++ engines/typescript-obfuscator/tsconfig.json
@@ -1,6 +1,10 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
+        "module": "CommonJS",
+        "target": "ES2020",
+        "declaration": true,
+        "emitDeclarationOnly": false,
         "outDir": "dist",
         "rootDir": "src",
         "baseUrl": "."
