+++ example/.gitignore
@@ -1,0 +1,1 @@
+**/.declassify-maps
