+++ libs/shared-config/.gitignore
@@ -1,0 +1,1 @@
+**/declassify-shared-config-*.tgz
