+++ libs/shared-config/package.json
@@ -4,6 +4,9 @@
     "version": "0.1.0",
     "main": "dist/index.js",
     "types": "dist/index.d.ts",
+    "files": [
+        "dist"
+    ],
     "scripts": {
         "build": "tsc -p ."
     },
