+++ libs/shared-config/tsconfig.json
@@ -1,6 +1,8 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
+        "module": "CommonJS",
+        "target": "ES2020",
         "outDir": "dist",
         "rootDir": "src",
         "baseUrl": "."
